num_workers=-1 (#552)
* sgd merged with nlp refactoring


* output files removed


* remove old code


* black


* clean up


* import fix


* eval callback for training data added


* clean up


* clean up


* clean up


* clean up


* clean up


* clean up


* dropout to token_emb added


* loss debug


* del


* clean up


* clean up


* updated lp


* rebase


* fix after rebase


* loss reduction=sum, F.gelu


* shuffle on


* loss log update


* loss explicit dim


* clean up


* unittest fix


* fix


* black


* clean up


* fix


* test


* schema


* add emb_dim as param


* ex_id_num added


* multi gpu support added


* multi gpu fix


* fix


* Fixed signing.


* Added readthedocs.yml.


* Added readthedocs.yml.

+++ nemo/collections/nlp/callbacks/sgd_callback.py
@@ -69,7 +69,9 @@ def eval_iter_callback(tensors, global_vars, ids_to_service_names_dict, eval_dat
     start_idx = torch.arange(max_num_tokens, device=device).view(1, 1, -1, 1)
     end_idx = torch.arange(max_num_tokens, device=device).view(1, 1, 1, -1)
     invalid_index_mask = (start_idx > end_idx).repeat(batch_size, max_num_noncat_slots, 1, 1)
-    total_scores = torch.where(invalid_index_mask, torch.zeros(total_scores.size(), device=device), total_scores)
+    total_scores = torch.where(
+        invalid_index_mask, torch.zeros(total_scores.size(), device=device, dtype=total_scores.dtype), total_scores
+    )
     max_span_index = torch.argmax(total_scores.view(-1, max_num_noncat_slots, max_num_tokens ** 2), axis=-1)
     span_start_index = torch.div(max_span_index, max_num_tokens)
     span_end_index = torch.fmod(max_span_index, max_num_tokens)
@@ -133,9 +135,9 @@ def evaluate(prediction_dir, data_dir, eval_dataset, output_metric_file):
     # Write the aggregated metrics values.
     with open(output_metric_file, "w") as f:
         json.dump(all_metric_aggregate, f, indent=2, separators=(",", ": "), sort_keys=True)
-    f.close()
+        f.close()
     # Write the per-frame metrics values with the corrresponding dialogue frames.
     with open(os.path.join(prediction_dir, PER_FRAME_OUTPUT_FILENAME), "w") as f:
         json.dump(dataset_hyp, f, indent=2, separators=(",", ": "))
-    f.close()
+        f.close()
     return all_metric_aggregate[ALL_SERVICES]
